Added flex and inset shadow in library.html
The main div containing the content and hero now uses flex for the following fixes to work.
The main div now has h-screen as a minimum, so that no whitespace is below the site in case there is no content.
The content div will fill the available space, which will fill the screen at minimum.
Added custom inset shadow.