Modified `server.js` to include a base URL
`server.js` now looks for "tailwind-infinix-website" at the start of the URL. If it detects it, it passes through a trimmed version of the URL removing "tailwind-infinix-website". This makes it so that the URL used is similar for the Github Pages when developing.
Modified `router.js` to adapt to the changes above.